requirements.txt
customtkinter
pypdf2
pymupdf
tkPDFViewer
pillow
pdf2image